+++ fidget/src/core/compiler/alloc.rs
@@ -251,6 +251,9 @@ impl<const N: usize> RegisterAllocator<N> {
             SsaOp::RecipReg(out, arg) => (out, arg, RegOp::RecipReg),
             SsaOp::SqrtReg(out, arg) => (out, arg, RegOp::SqrtReg),
             SsaOp::SquareReg(out, arg) => (out, arg, RegOp::SquareReg),
+            SsaOp::FloorReg(out, arg) => (out, arg, RegOp::FloorReg),
+            SsaOp::CeilReg(out, arg) => (out, arg, RegOp::CeilReg),
+            SsaOp::RoundReg(out, arg) => (out, arg, RegOp::RoundReg),
             SsaOp::SinReg(out, arg) => (out, arg, RegOp::SinReg),
             SsaOp::CosReg(out, arg) => (out, arg, RegOp::CosReg),
             SsaOp::TanReg(out, arg) => (out, arg, RegOp::TanReg),
@@ -278,6 +281,9 @@ impl<const N: usize> RegisterAllocator<N> {
             | SsaOp::RecipReg(..)
             | SsaOp::SqrtReg(..)
             | SsaOp::SquareReg(..)
+            | SsaOp::FloorReg(..)
+            | SsaOp::CeilReg(..)
+            | SsaOp::RoundReg(..)
             | SsaOp::CopyReg(..)
             | SsaOp::SinReg(..)
             | SsaOp::CosReg(..)
@@ -295,6 +301,8 @@ impl<const N: usize> RegisterAllocator<N> {
             | SsaOp::MulRegImm(..)
             | SsaOp::DivRegImm(..)
             | SsaOp::DivImmReg(..)
+            | SsaOp::AtanImmReg(..)
+            | SsaOp::AtanRegImm(..)
             | SsaOp::MinRegImm(..)
             | SsaOp::MaxRegImm(..)
             | SsaOp::CompareRegImm(..)
@@ -308,6 +316,7 @@ impl<const N: usize> RegisterAllocator<N> {
             | SsaOp::SubRegReg(..)
             | SsaOp::MulRegReg(..)
             | SsaOp::DivRegReg(..)
+            | SsaOp::AtanRegReg(..)
             | SsaOp::MinRegReg(..)
             | SsaOp::MaxRegReg(..)
             | SsaOp::CompareRegReg(..)
@@ -490,6 +499,9 @@ impl<const N: usize> RegisterAllocator<N> {
             SsaOp::DivRegReg(out, lhs, rhs) => {
                 (out, lhs, rhs, RegOp::DivRegReg)
             }
+            SsaOp::AtanRegReg(out, lhs, rhs) => {
+                (out, lhs, rhs, RegOp::AtanRegReg)
+            }
             SsaOp::MinRegReg(out, lhs, rhs) => {
                 (out, lhs, rhs, RegOp::MinRegReg)
             }
@@ -611,6 +623,12 @@ impl<const N: usize> RegisterAllocator<N> {
             }
             SsaOp::DivImmReg(out, arg, imm) => {
                 (out, arg, imm, RegOp::DivImmReg)
+            }
+            SsaOp::AtanRegImm(out, arg, imm) => {
+                (out, arg, imm, RegOp::AtanRegImm)
+            }
+            SsaOp::AtanImmReg(out, arg, imm) => {
+                (out, arg, imm, RegOp::AtanImmReg)
             }
             SsaOp::MinRegImm(out, arg, imm) => {
                 (out, arg, imm, RegOp::MinRegImm)