+++ fidget/src/core/compiler/ssa_tape.rs
@@ -129,6 +129,11 @@ impl SsaTape {
                             SsaOp::DivRegImm,
                             SsaOp::DivImmReg,
                         ),
+                        BinaryOpcode::Atan => (
+                            SsaOp::AtanRegReg,
+                            SsaOp::AtanRegImm,
+                            SsaOp::AtanImmReg,
+                        ),
                         BinaryOpcode::Min => (
                             SsaOp::MinRegReg,
                             SsaOp::MinRegImm,
@@ -201,6 +206,9 @@ impl SsaTape {
                         UnaryOpcode::Recip => SsaOp::RecipReg,
                         UnaryOpcode::Sqrt => SsaOp::SqrtReg,
                         UnaryOpcode::Square => SsaOp::SquareReg,
+                        UnaryOpcode::Floor => SsaOp::FloorReg,
+                        UnaryOpcode::Ceil => SsaOp::CeilReg,
+                        UnaryOpcode::Round => SsaOp::RoundReg,
                         UnaryOpcode::Sin => SsaOp::SinReg,
                         UnaryOpcode::Cos => SsaOp::CosReg,
                         UnaryOpcode::Tan => SsaOp::TanReg,
@@ -262,6 +270,9 @@ impl SsaTape {
                 | SsaOp::SqrtReg(out, arg)
                 | SsaOp::CopyReg(out, arg)
                 | SsaOp::SquareReg(out, arg)
+                | SsaOp::FloorReg(out, arg)
+                | SsaOp::CeilReg(out, arg)
+                | SsaOp::RoundReg(out, arg)
                 | SsaOp::SinReg(out, arg)
                 | SsaOp::CosReg(out, arg)
                 | SsaOp::TanReg(out, arg)
@@ -277,6 +288,9 @@ impl SsaTape {
                         SsaOp::RecipReg(..) => "RECIP",
                         SsaOp::SqrtReg(..) => "SQRT",
                         SsaOp::SquareReg(..) => "SQUARE",
+                        SsaOp::FloorReg(..) => "FLOOR",
+                        SsaOp::CeilReg(..) => "CEIL",
+                        SsaOp::RoundReg(..) => "ROUND",
                         SsaOp::SinReg(..) => "SIN",
                         SsaOp::CosReg(..) => "COS",
                         SsaOp::TanReg(..) => "TAN",
@@ -300,11 +314,13 @@ impl SsaTape {
                 | SsaOp::MaxRegReg(out, lhs, rhs)
                 | SsaOp::ModRegReg(out, lhs, rhs)
                 | SsaOp::AndRegReg(out, lhs, rhs)
+                | SsaOp::AtanRegReg(out, lhs, rhs)
                 | SsaOp::OrRegReg(out, lhs, rhs) => {
                     let op = match op {
                         SsaOp::AddRegReg(..) => "ADD",
                         SsaOp::MulRegReg(..) => "MUL",
                         SsaOp::DivRegReg(..) => "DIV",
+                        SsaOp::AtanRegReg(..) => "ATAN",
                         SsaOp::SubRegReg(..) => "SUB",
                         SsaOp::MinRegReg(..) => "MIN",
                         SsaOp::MaxRegReg(..) => "MAX",
@@ -322,6 +338,8 @@ impl SsaTape {
                 | SsaOp::DivImmReg(out, arg, imm)
                 | SsaOp::SubImmReg(out, arg, imm)
                 | SsaOp::SubRegImm(out, arg, imm)
+                | SsaOp::AtanRegImm(out, arg, imm)
+                | SsaOp::AtanImmReg(out, arg, imm)
                 | SsaOp::MinRegImm(out, arg, imm)
                 | SsaOp::MaxRegImm(out, arg, imm)
                 | SsaOp::ModRegImm(out, arg, imm)
@@ -335,6 +353,8 @@ impl SsaTape {
                         SsaOp::DivRegImm(..) => ("DIV", false),
                         SsaOp::SubImmReg(..) => ("SUB", true),
                         SsaOp::SubRegImm(..) => ("SUB", false),
+                        SsaOp::AtanImmReg(..) => ("ATAN", true),
+                        SsaOp::AtanRegImm(..) => ("ATAN", false),
                         SsaOp::MinRegImm(..) => ("MIN", false),
                         SsaOp::MaxRegImm(..) => ("MAX", false),
                         SsaOp::ModRegImm(..) => ("MOD", false),