+++ fidget/src/core/context/mod.rs
@@ -549,6 +549,49 @@ impl Context {
         self.op_unary(a, UnaryOpcode::Square)
     }
 
+    /// Builds a node which takes the floor of its input
+    /// ```
+    /// # let mut ctx = fidget::context::Context::new();
+    /// let x = ctx.x();
+    /// let op = ctx.floor(x).unwrap();
+    /// let v = ctx.eval_xyz(op, 1.2, 0.0, 0.0).unwrap();
+    /// assert_eq!(v, 1.0);
+    /// ```
+    pub fn floor<A: IntoNode>(&mut self, a: A) -> Result<Node, Error> {
+        let a = a.into_node(self)?;
+        self.op_unary(a, UnaryOpcode::Floor)
+    }
+
+    /// Builds a node which takes the ceiling of its input
+    /// ```
+    /// # let mut ctx = fidget::context::Context::new();
+    /// let x = ctx.x();
+    /// let op = ctx.ceil(x).unwrap();
+    /// let v = ctx.eval_xyz(op, 1.2, 0.0, 0.0).unwrap();
+    /// assert_eq!(v, 2.0);
+    /// ```
+    pub fn ceil<A: IntoNode>(&mut self, a: A) -> Result<Node, Error> {
+        let a = a.into_node(self)?;
+        self.op_unary(a, UnaryOpcode::Ceil)
+    }
+
+    /// Builds a node which rounds its input to the nearest integer
+    /// ```
+    /// # let mut ctx = fidget::context::Context::new();
+    /// let x = ctx.x();
+    /// let op = ctx.round(x).unwrap();
+    /// let v = ctx.eval_xyz(op, 1.2, 0.0, 0.0).unwrap();
+    /// assert_eq!(v, 1.0);
+    /// let v = ctx.eval_xyz(op, 1.6, 0.0, 0.0).unwrap();
+    /// assert_eq!(v, 2.0);
+    /// let v = ctx.eval_xyz(op, 1.5, 0.0, 0.0).unwrap();
+    /// assert_eq!(v, 2.0); // rounds away from 0.0 if ambiguous
+    /// ```
+    pub fn round<A: IntoNode>(&mut self, a: A) -> Result<Node, Error> {
+        let a = a.into_node(self)?;
+        self.op_unary(a, UnaryOpcode::Round)
+    }
+
     /// Builds a node which performs subtraction.
     /// ```
     /// # let mut ctx = fidget::context::Context::new();
@@ -573,8 +616,7 @@ impl Context {
         }
     }
 
-    /// Builds a node which performs division. Under the hood, `a / b` is
-    /// converted into `a * (1 / b)`.
+    /// Builds a node which performs division.
     /// ```
     /// # let mut ctx = fidget::context::Context::new();
     /// let x = ctx.x();
@@ -596,6 +638,26 @@ impl Context {
             (_, Some(one)) if one == 1.0 => Ok(a),
             _ => self.op_binary(a, b, BinaryOpcode::Div),
         }
+    }
+
+    /// Builds a node which computes `atan2(y, x)`
+    /// ```
+    /// # let mut ctx = fidget::context::Context::new();
+    /// let x = ctx.x();
+    /// let y = ctx.y();
+    /// let op = ctx.atan2(y, x).unwrap();
+    /// let v = ctx.eval_xyz(op, 0.0, 1.0, 0.0).unwrap();
+    /// assert_eq!(v, std::f64::consts::FRAC_PI_2);
+    /// ```
+    pub fn atan2<A: IntoNode, B: IntoNode>(
+        &mut self,
+        y: A,
+        x: B,
+    ) -> Result<Node, Error> {
+        let y = y.into_node(self)?;
+        let x = x.into_node(self)?;
+
+        self.op_binary(y, x, BinaryOpcode::Atan)
     }
 
     /// Builds a node that compares two values
@@ -738,6 +800,7 @@ impl Context {
                     BinaryOpcode::Sub => a - b,
                     BinaryOpcode::Mul => a * b,
                     BinaryOpcode::Div => a / b,
+                    BinaryOpcode::Atan => a.atan2(b),
                     BinaryOpcode::Min => a.min(b),
                     BinaryOpcode::Max => a.max(b),
                     BinaryOpcode::Compare => a
@@ -771,6 +834,9 @@ impl Context {
                     UnaryOpcode::Recip => 1.0 / a,
                     UnaryOpcode::Sqrt => a.sqrt(),
                     UnaryOpcode::Square => a * a,
+                    UnaryOpcode::Floor => a.floor(),
+                    UnaryOpcode::Ceil => a.ceil(),
+                    UnaryOpcode::Round => a.round(),
                     UnaryOpcode::Sin => a.sin(),
                     UnaryOpcode::Cos => a.cos(),
                     UnaryOpcode::Tan => a.tan(),
@@ -837,6 +903,9 @@ impl Context {
                 "neg" => ctx.neg(pop()?)?,
                 "sqrt" => ctx.sqrt(pop()?)?,
                 "square" => ctx.square(pop()?)?,
+                "floor" => ctx.floor(pop()?)?,
+                "ceil" => ctx.ceil(pop()?)?,
+                "round" => ctx.round(pop()?)?,
                 "sin" => ctx.sin(pop()?)?,
                 "cos" => ctx.cos(pop()?)?,
                 "tan" => ctx.tan(pop()?)?,
@@ -851,6 +920,7 @@ impl Context {
                 "min" => ctx.min(pop()?, pop()?)?,
                 "max" => ctx.max(pop()?, pop()?)?,
                 "div" => ctx.div(pop()?, pop()?)?,
+                "atan2" => ctx.atan2(pop()?, pop()?)?,
                 "sub" => ctx.sub(pop()?, pop()?)?,
                 "compare" => ctx.compare(pop()?, pop()?)?,
                 "mod" => ctx.modulo(pop()?, pop()?)?,
@@ -897,6 +967,7 @@ impl Context {
                 BinaryOpcode::Sub => out += "sub",
                 BinaryOpcode::Mul => out += "mul",
                 BinaryOpcode::Div => out += "div",
+                BinaryOpcode::Atan => out += "atan2",
                 BinaryOpcode::Min => out += "min",
                 BinaryOpcode::Max => out += "max",
                 BinaryOpcode::Compare => out += "compare",
@@ -910,6 +981,9 @@ impl Context {
                 UnaryOpcode::Recip => out += "recip",
                 UnaryOpcode::Sqrt => out += "sqrt",
                 UnaryOpcode::Square => out += "square",
+                UnaryOpcode::Floor => out += "floor",
+                UnaryOpcode::Ceil => out += "ceil",
+                UnaryOpcode::Round => out += "round",
                 UnaryOpcode::Sin => out += "sin",
                 UnaryOpcode::Cos => out += "cos",
                 UnaryOpcode::Tan => out += "tan",