+++ fidget/src/core/eval/test/mod.rs
@@ -201,6 +201,9 @@ pub mod canonical {
     declare_canonical_unary!(Context::ln, |a| a.ln());
     declare_canonical_unary!(Context::square, |a| a * a);
     declare_canonical_unary!(Context::sqrt, |a| a.sqrt());
+    declare_canonical_unary!(Context::floor, |a| a.floor());
+    declare_canonical_unary!(Context::ceil, |a| a.ceil());
+    declare_canonical_unary!(Context::round, |a| a.round());
     declare_canonical_unary!(Context::not, |a| (a == 0.0).into(), |a| a == 0.0);
 
     declare_canonical_binary!(Context::add, |a, b| a + b);
@@ -259,6 +262,7 @@ pub mod canonical {
         |a, b| if a != 0.0 { a } else { b },
         |a, _b| a == 0.0 // discontinuity, because either side snaps to a
     );
+    declare_canonical_binary!(Context::atan2, |y, x| y.atan2(x));
 }
 
 #[macro_export]
@@ -297,6 +301,9 @@ macro_rules! all_unary_tests {
         $crate::one_unary_test!($tester, ln);
         $crate::one_unary_test!($tester, not);
         $crate::one_unary_test!($tester, square);
+        $crate::one_unary_test!($tester, floor);
+        $crate::one_unary_test!($tester, ceil);
+        $crate::one_unary_test!($tester, round);
         $crate::one_unary_test!($tester, sqrt);
     };
 }
@@ -308,6 +315,7 @@ macro_rules! all_binary_tests {
         $crate::one_binary_test!($tester, sub);
         $crate::one_binary_test!($tester, mul);
         $crate::one_binary_test!($tester, div);
+        $crate::one_binary_test!($tester, atan2);
         $crate::one_binary_test!($tester, min);
         $crate::one_binary_test!($tester, max);
         $crate::one_binary_test!($tester, compare);