+++ fidget/src/core/vm/mod.rs
@@ -300,6 +300,15 @@ impl<const N: usize> TracingEvaluator for VmIntervalEval<N> {
                 RegOp::SquareReg(out, arg) => {
                     v[out] = v[arg].square();
                 }
+                RegOp::FloorReg(out, arg) => {
+                    v[out] = v[arg].floor();
+                }
+                RegOp::CeilReg(out, arg) => {
+                    v[out] = v[arg].ceil();
+                }
+                RegOp::RoundReg(out, arg) => {
+                    v[out] = v[arg].round();
+                }
                 RegOp::SinReg(out, arg) => {
                     v[out] = v[arg].sin();
                 }
@@ -346,6 +355,16 @@ impl<const N: usize> TracingEvaluator for VmIntervalEval<N> {
                 RegOp::DivImmReg(out, arg, imm) => {
                     let imm: Interval = imm.into();
                     v[out] = imm / v[arg];
+                }
+                RegOp::AtanRegImm(out, arg, imm) => {
+                    v[out] = v[arg].atan2(imm.into());
+                }
+                RegOp::AtanImmReg(out, arg, imm) => {
+                    let imm: Interval = imm.into();
+                    v[out] = imm.atan2(v[arg]);
+                }
+                RegOp::AtanRegReg(out, lhs, rhs) => {
+                    v[out] = v[lhs].atan2(v[rhs]);
                 }
                 RegOp::SubImmReg(out, arg, imm) => {
                     v[out] = Interval::from(imm) - v[arg];
@@ -521,6 +540,15 @@ impl<const N: usize> TracingEvaluator for VmPointEval<N> {
                     let s = v[arg];
                     v[out] = s * s;
                 }
+                RegOp::FloorReg(out, arg) => {
+                    v[out] = v[arg].floor();
+                }
+                RegOp::CeilReg(out, arg) => {
+                    v[out] = v[arg].ceil();
+                }
+                RegOp::RoundReg(out, arg) => {
+                    v[out] = v[arg].round();
+                }
                 RegOp::SinReg(out, arg) => {
                     v[out] = v[arg].sin();
                 }
@@ -560,6 +588,15 @@ impl<const N: usize> TracingEvaluator for VmPointEval<N> {
                 }
                 RegOp::DivImmReg(out, arg, imm) => {
                     v[out] = imm / v[arg];
+                }
+                RegOp::AtanRegImm(out, arg, imm) => {
+                    v[out] = v[arg].atan2(imm.into());
+                }
+                RegOp::AtanImmReg(out, arg, imm) => {
+                    v[out] = imm.atan2(v[arg]);
+                }
+                RegOp::AtanRegReg(out, lhs, rhs) => {
+                    v[out] = v[lhs].atan2(v[rhs]);
                 }
                 RegOp::SubImmReg(out, arg, imm) => {
                     v[out] = imm - v[arg];
@@ -836,6 +873,21 @@ impl<const N: usize> BulkEvaluator for VmFloatSliceEval<N> {
                         v[out][i] = s * s;
                     }
                 }
+                RegOp::FloorReg(out, arg) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].floor();
+                    }
+                }
+                RegOp::CeilReg(out, arg) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].ceil();
+                    }
+                }
+                RegOp::RoundReg(out, arg) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].round();
+                    }
+                }
                 RegOp::SinReg(out, arg) => {
                     for i in 0..size {
                         v[out][i] = v[arg][i].sin();
@@ -904,6 +956,21 @@ impl<const N: usize> BulkEvaluator for VmFloatSliceEval<N> {
                 RegOp::DivImmReg(out, arg, imm) => {
                     for i in 0..size {
                         v[out][i] = imm / v[arg][i];
+                    }
+                }
+                RegOp::AtanRegImm(out, arg, imm) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].atan2(imm);
+                    }
+                }
+                RegOp::AtanImmReg(out, arg, imm) => {
+                    for i in 0..size {
+                        v[out][i] = imm.atan2(v[arg][i]);
+                    }
+                }
+                RegOp::AtanRegReg(out, lhs, rhs) => {
+                    for i in 0..size {
+                        v[out][i] = v[lhs][i].atan2(v[rhs][i]);
                     }
                 }
                 RegOp::SubImmReg(out, arg, imm) => {
@@ -1126,6 +1193,21 @@ impl<const N: usize> BulkEvaluator for VmGradSliceEval<N> {
                         v[out][i] = s * s;
                     }
                 }
+                RegOp::FloorReg(out, arg) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].floor();
+                    }
+                }
+                RegOp::CeilReg(out, arg) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].ceil();
+                    }
+                }
+                RegOp::RoundReg(out, arg) => {
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].round();
+                    }
+                }
                 RegOp::SinReg(out, arg) => {
                     for i in 0..size {
                         v[out][i] = v[arg][i].sin();
@@ -1192,9 +1274,26 @@ impl<const N: usize> BulkEvaluator for VmGradSliceEval<N> {
                     }
                 }
                 RegOp::DivImmReg(out, arg, imm) => {
-                    let imm: Grad = imm.into();
+                    let imm = Grad::from(imm);
                     for i in 0..size {
                         v[out][i] = imm / v[arg][i];
+                    }
+                }
+                RegOp::AtanRegImm(out, arg, imm) => {
+                    let imm = Grad::from(imm);
+                    for i in 0..size {
+                        v[out][i] = v[arg][i].atan2(imm);
+                    }
+                }
+                RegOp::AtanImmReg(out, arg, imm) => {
+                    let imm = Grad::from(imm);
+                    for i in 0..size {
+                        v[out][i] = imm.atan2(v[arg][i]);
+                    }
+                }
+                RegOp::AtanRegReg(out, lhs, rhs) => {
+                    for i in 0..size {
+                        v[out][i] = v[lhs][i].atan2(v[rhs][i]);
                     }
                 }
                 RegOp::SubImmReg(out, arg, imm) => {