+++ fidget/src/jit/x86_64/point.rs
@@ -177,6 +177,27 @@ impl Assembler for PointAssembler {
             ; vmulss Rx(reg(out_reg)), Rx(reg(lhs_reg)), Rx(reg(lhs_reg))
         );
     }
+
+    // TODO optimize these three functions
+    fn build_floor(&mut self, out_reg: u8, lhs_reg: u8) {
+        extern "sysv64" fn float_floor(f: f32) -> f32 {
+            f.floor()
+        }
+        self.call_fn_unary(out_reg, lhs_reg, float_floor);
+    }
+    fn build_ceil(&mut self, out_reg: u8, lhs_reg: u8) {
+        extern "sysv64" fn float_ceil(f: f32) -> f32 {
+            f.ceil()
+        }
+        self.call_fn_unary(out_reg, lhs_reg, float_ceil);
+    }
+    fn build_round(&mut self, out_reg: u8, lhs_reg: u8) {
+        extern "sysv64" fn float_round(f: f32) -> f32 {
+            f.round()
+        }
+        self.call_fn_unary(out_reg, lhs_reg, float_round);
+    }
+
     fn build_add(&mut self, out_reg: u8, lhs_reg: u8, rhs_reg: u8) {
         dynasm!(self.0.ops
             ; vaddss Rx(reg(out_reg)), Rx(reg(lhs_reg)), Rx(reg(rhs_reg))
@@ -196,6 +217,12 @@ impl Assembler for PointAssembler {
         dynasm!(self.0.ops
             ; vdivss Rx(reg(out_reg)), Rx(reg(lhs_reg)), Rx(reg(rhs_reg))
         );
+    }
+    fn build_atan2(&mut self, out_reg: u8, lhs_reg: u8, rhs_reg: u8) {
+        extern "sysv64" fn float_atan2(y: f32, x: f32) -> f32 {
+            y.atan2(x)
+        }
+        self.call_fn_binary(out_reg, lhs_reg, rhs_reg, float_atan2);
     }
     fn build_max(&mut self, out_reg: u8, lhs_reg: u8, rhs_reg: u8) {
         dynasm!(self.0.ops
@@ -430,6 +457,73 @@ impl PointAssembler {
 
             // call the function
             ; movss xmm0, Rx(reg(arg_reg))
+            ; mov rsi, QWORD addr as _
+            ; call rsi
+
+            // Restore float registers
+            ; movss xmm4, [rsp]
+            ; movss xmm5, [rsp + 0x4]
+            ; movss xmm6, [rsp + 0x8]
+            ; movss xmm7, [rsp + 0xc]
+            ; movss xmm8, [rsp + 0x10]
+            ; movss xmm9, [rsp + 0x14]
+            ; movss xmm10, [rsp + 0x18]
+            ; movss xmm11, [rsp + 0x1c]
+            ; movss xmm12, [rsp + 0x20]
+            ; movss xmm13, [rsp + 0x24]
+            ; movss xmm14, [rsp + 0x28]
+            ; movss xmm15, [rsp + 0x2c]
+
+            // Restore pointers
+            ; mov rdi, r12
+            ; mov rsi, r13
+            ; mov rdx, r14
+
+            ; movss Rx(reg(out_reg)), xmm0
+        );
+    }
+    fn call_fn_binary(
+        &mut self,
+        out_reg: u8,
+        lhs_reg: u8,
+        rhs_reg: u8,
+        f: extern "sysv64" fn(f32, f32) -> f32,
+    ) {
+        // Back up a few callee-saved registers that we're about to use
+        if !self.0.saved_callee_regs {
+            dynasm!(self.0.ops
+                ; mov [rbp - 0x8], r12
+                ; mov [rbp - 0x10], r13
+                ; mov [rbp - 0x18], r14
+            );
+            self.0.saved_callee_regs = true
+        }
+        let addr = f as usize;
+        dynasm!(self.0.ops
+            // Back up pointers to caller-saved registers
+            ; mov r12, rdi
+            ; mov r13, rsi
+            ; mov r14, rdx
+
+            // Back up all register values to the stack
+            ; movss [rsp], xmm4
+            ; movss [rsp + 0x4], xmm5
+            ; movss [rsp + 0x8], xmm6
+            ; movss [rsp + 0xc], xmm7
+            ; movss [rsp + 0x10], xmm8
+            ; movss [rsp + 0x14], xmm9
+            ; movss [rsp + 0x18], xmm10
+            ; movss [rsp + 0x1c], xmm11
+            ; movss [rsp + 0x20], xmm12
+            ; movss [rsp + 0x24], xmm13
+            ; movss [rsp + 0x28], xmm14
+            ; movss [rsp + 0x2c], xmm15
+
+            // call the function.  Note the ordering here: xmm0 could be LHS /
+            // RHS if we're doing a call with an immediate, so we overwrite it
+            // last.
+            ; movss xmm1, Rx(reg(rhs_reg))
+            ; movss xmm0, Rx(reg(lhs_reg))
             ; mov rsi, QWORD addr as _
             ; call rsi